requirements.txt
asgiref==3.3.1
Django==3.1.6
pytz==2021.1
sqlparse==0.4.1
PyMySQL==1.0.2
mistune~=0.8.4
django-ckeditor~=6.0.0
djangorestframework~=3.12.4